modified:   opt_crm.c Added two fields to arterial_desc_t for setting the normal patterns that San Jose has for their arterials. They change patterns at 7 and 9 AM, and we're controlling during that time. Changed -d option from debug to global disable Added initialization of db_set_pattern local variable. Added time check for control period Added reset of pattern to San Jose pattern when occupancy drops blow 30% Added global disable flag status to log file Added fflush to log file pointer
modified:   resource.c
In occupancy_aggregation_onramp_queue, set controller_data->num_meter, confidence->num_total_vals, confidence->num_good_vals, and MAX_QUEUE_LOOPS to 1.
This was done bcause in this corridor, the diagonal and loop onramps are considered to be two lanes of the same onramp inside the controller. The backup
was occurring only on the diagonal onramp (lane1), so averaging the two gave a fals reading for occupancy.